# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cirvee Referral System API",
        "description": "Production-grade referral system API with JWT authentication",
        "version": "1.0.0",
        "contact": {
            "name": "Cirvee",
            "email": "admin@cirvee.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Development server"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "DashboardStats": {
                "type": "object",
                "properties": {
                    "total_earnings": {
                        "type": "integer"
                    },
                    "pending_balance": {
                        "type": "integer"
                    },
                    "total_referrals": {
                        "type": "integer"
                    },
                    "total_clicks": {
                        "type": "integer"
                    },
                    "total_payouts": {
                        "type": "integer"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/api/v1/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard stats",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardStats"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/admin/referrals": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all referrals",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of referrals"
                    }
                }
            }
        },
        "/api/v1/admin/students": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all students/users",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                }
            }
        },
        "/api/v1/admin/payouts": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all payouts",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "approved",
                                "rejected"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of payouts"
                    }
                }
            }
        },
        "/api/v1/admin/payouts/{id}": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update payout status",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "approved",
                                            "rejected"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                }
            }
        },
...
    }
}